Rework the percent-encoding support (closes #1)
The rules for percent-encoding are a bit complex, so the internal
interface makes the specific state of the string components more
abstract. There is a slight run-time cost to this, but not a
significant one.

The biggest noticable difference is that the external interface
to the library is different now, with Uri.t being abstract and
accessor functions required to access the internals of a Uri.

It's actually easier to use this way...