Allow configuration of query encoding in uri's
Uri.t can now be created with an optional `query_scheme` parameter that
controls the encoding of the query string. If it's omitted then the
encoding defaults to the uri's own scheme.

An implementation is given for query encoding according to amazon's
standards to be used when writing clients for AWS.